[ refactor ] 패키지 변경 및 @Data 어노테이션 제거
