+++ module-core/src/main/java/dto/internal/GenreDTO.java
@@ -1,8 +1,5 @@
-package dto;
+package dto.internal;
 
-import lombok.Data;
-
-@Data
 public class GenreDTO {
 	private Long id;
 	private String name;
