+++ module-core/src/main/java/dto/internal/RatingDTO.java
@@ -1,8 +1,5 @@
-package dto;
+package dto.internal;
 
-import lombok.Data;
-
-@Data
 public class RatingDTO {
 	private Long id;
 	private String name;
