+++ module-core/src/main/java/dto/internal/ScreenDTO.java
@@ -1,8 +1,5 @@
-package dto;
+package dto.internal;
 
-import lombok.Data;
-
-@Data
 public class ScreenDTO {
 	private Long screen_id;
 	private String name;
